backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask-cors>=5.0.0",
    "flask-migrate>=4.0.7",
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]
